dashboard-operacoes/requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.23.0
python-dotenv==1.0.1
numpy==2.0.1
supabase==2.5.5
pydantic==2.8.2
